Move styled components into their respective directories
