+++ react-playground/src/components/Menu/Menu.styles.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const MenuStyled = styled.ul`
+  display: flex;
+  flex-direction: column;
+  padding: 16px;
+  position: absolute;
+  top: 60px;
+  right: 30px;
+  background: #ffffff;
+  box-shadow: 0px 4px 20px rgba(0, 0, 0, 0.1),
+    0px 2px 0px -10px rgba(27, 33, 50, 0.08);
+  border-radius: 8px;
+`;
