+++ react-playground/src/components/MenuItem/MenuItem.styles.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const MenuItemStyled = styled.li`
+  list-style: none;
+  color: #1a1a19;
+  font-family: "Poppins", sans-serif;
+  font-style: normal;
+  font-size: 18px;
+  line-height: 33px;
+  flex: none;
+  order: 1;
+  flex-grow: 0;
+  margin: 5px 45px 0px 15px;
+`;
